% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkchoc.R
\docType{data}
\name{milkchoc}
\alias{milkchoc}
\title{Check-All-That-Apply data}
\format{
An object of class \code{data.frame} with 280 rows and 8 columns.
}
\usage{
data(milkchoc)
}
\description{
Data coming from a Check-All-That-Apply experiment on milk chocolates. This dataset corresponds to the flavor data used as examples and described more precisely in Mahieu, Schlich, Visalli, and Cardot (2021)
}
\examples{
data(milkchoc)
}
\references{
Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2021). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Food Quality and Preference, 93.
}
\keyword{datasets}
