% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.dimensionality.test.R
\name{mr.dimensionality.test}
\alias{mr.dimensionality.test}
\title{Multiple-response dimensionality test}
\usage{
mr.dimensionality.test(data, nperm = 2000, alpha = 0.05, ncores = 2)
}
\arguments{
\item{data}{A data.frame of observations in rows whose first column is a factor (the categories) and subsequent columns are binary numeric or integer, each column being a response option}

\item{nperm}{Number of permuted datasets to estimate the distribution of the statistic under the null hypothesis. See details}

\item{alpha}{The α risk of the test}

\item{ncores}{Number of cores used to estimate the null distribution. Default is 2. See details}
}
\value{
A list with the following elements:
\describe{
\item{dim.sig}{The number of significant dimensions}
\item{statistics}{Observed multiple-response chi-square statistic of each dimension}
\item{p.values}{P-value of the test of each dimension adjusted for closed testing procedure}
}
}
\description{
Performs a multiple-response dimensionality test as defined in Mahieu, Schlich, Visalli, and Cardot (2020) using random permutations to estimate the null distribution. This test results of the transposition of the dimensionality test introduced in Mahieu, Visalli, and Schlich (2020) from the usual chi-square framework to the modified chi-square framework introduced in Loughin and Scherer (1998)
}
\details{
\itemize{
\item \strong{nperm}: The distribution of the statistic under the null hypothesis of no associations between categories and response options is estimated using \emph{nperm} datasets generated thanks to random permutations of the response vectors along observations.
\item \strong{ncores}: The more cores are added in the process, the faster the results will be obtained. The number of available cores is accessible using \code{\link[parallel]{detectCores}}. The parallel tasks are closed once the \emph{nperm} datasets are generated.
}
}
\examples{
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

parallel::detectCores()

mr.dimensionality.test(dset)

}
\references{
Loughin, T. M., & Scherer, P. N. (1998). Testing for Association in Contingency Tables with Multiple Column Responses. Biometrics, 54(2), 630-637.

Mahieu, B., Visalli, M., & Schlich, P. (2020). Accounting for the dimensionality of the dependence in analyses of contingency tables obtained with Check-All-That-Apply and Free-Comment. Food Quality and Preference, 83.

Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2020). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Manuscript submitted for publication.
}
