% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiResponseR-package.R
\docType{package}
\name{MultiResponseR-package}
\alias{MultiResponseR}
\alias{MultiResponseR-package}
\title{MultiResponseR: Analysis of Data from Multiple-Response Questionnaires}
\description{
Implements the multiple-response chi-square framework introduced in Mahieu, Schlich, Visalli, and Cardot (2021).
}
\references{
Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2021). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Food Quality and Preference, 93.

Loughin, T. M., & Scherer, P. N. (1998). Testing for Association in Contingency Tables with Multiple Column Responses. Biometrics, 54(2), 630-637.
}
\seealso{
\code{\link[MultiResponseR]{mr.chisq.test}}

\code{\link[MultiResponseR]{mr.dimensionality.test}}

\code{\link[MultiResponseR]{mr.sig.cell}}

\code{\link[MultiResponseR]{mrCA}}

\code{\link[MultiResponseR]{plt.mr.sig.cell}}

\code{\link[MultiResponseR]{plt.mrCA}}

\code{\link[MultiResponseR]{sensory.mr.dimensionality.test}}

\code{\link[MultiResponseR]{sensory.mr.sig.cell}}

\code{\link[MultiResponseR]{sensory.mrCA}}

\code{\link[MultiResponseR]{sensory.overall.analysis}}
}
\author{
\strong{Maintainer}: Benjamin Mahieu \email{benjamin.mahieu@oniris-nantes.fr}

}
\keyword{internal}
