% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mr.sig.cell.R
\name{plot.mr.sig.cell}
\alias{plot.mr.sig.cell}
\title{Plot significant cells}
\usage{
\method{plot}{mr.sig.cell}(
  x,
  alpha.1 = 0.05,
  alpha.2 = 0,
  choice = "percent.derived.cont",
  col.greater.1 = "green3",
  col.lower.1 = "orangered",
  col.greater.2 = "lightgreen",
  col.lower.2 = "lightsalmon",
  ...
)
}
\arguments{
\item{x}{A list returned by \code{\link[MultiResponseR]{mr.sig.cell}}}

\item{alpha.1}{The alpha risk to consider the tests as significant}

\item{alpha.2}{The alpha risk to consider the tests as showing a trend. If trends are not to be considered, \emph{alpha.2} can be set to 0 (Default)}

\item{choice}{Which table from \emph{res} should be plotted? Default is percent.derived.cont}

\item{col.greater.1}{The color used to highlight significant positive associations}

\item{col.lower.1}{The color used to highlight significant negative associations}

\item{col.greater.2}{The color used to highlight positive associations showing a trend}

\item{col.lower.2}{The color used to highlight negative associations showing a trend}

\item{...}{further arguments passed to or from other methods}
}
\value{
A table with cells highlighted
}
\description{
This function plots the results coming from \code{\link[MultiResponseR]{mr.sig.cell}}
}
\examples{
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

res=mr.sig.cell(dset)

plot(res)
}
