% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mrCA.R
\name{plot.mrCA}
\alias{plot.mrCA}
\title{Plot factor plan resulting from multiple-response Correspondence Analysis (MR-CA)}
\usage{
\method{plot}{mrCA}(
  x,
  axes = c(1, 2),
  alpha.total.bootstrap.test = 0.05,
  alpha.ellipse = alpha.total.bootstrap.test,
  select.rep = rownames(x$col.coord),
  rev.x = FALSE,
  rev.y = FALSE,
  size.points = 3.5,
  size.lab = 6,
  expansion = 1.25,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{A list returned by \code{\link[MultiResponseR]{mrCA}}}

\item{axes}{Which dimensions of the MR-CA should be plotted?}

\item{alpha.total.bootstrap.test}{The alpha risk of the total bootstrap tests. Only useful if the MR-CA was computed using \code{\link[MultiResponseR]{mrCA}} and ellipse=TRUE. See details}

\item{alpha.ellipse}{The alpha risk of the confidence ellipses. Only useful if the MR-CA was computed using \code{\link[MultiResponseR]{mrCA}} and ellipse=TRUE}

\item{select.rep}{A character vector specifying the response options to plot. By default, all response options are plotted}

\item{rev.x}{Should the horizontal plotted dimension be reversed? Useful in case of map comparisons to align categories}

\item{rev.y}{Should the vertical plotted dimension be reversed? Useful in case of map comparisons to align categories}

\item{size.points}{The size of the points used to represent the categories on the map}

\item{size.lab}{The size of the label on the map}

\item{expansion}{The factor of expansion applied to response options coordinates to increase readability}

\item{title}{An optional title to be added to the plot}

\item{...}{further arguments passed to or from other methods}
}
\value{
A MR-CA factor map
}
\description{
This function plots the results coming from \code{\link[MultiResponseR]{mrCA}}
}
\details{
\itemize{
\item \strong{alpha.total.bootstrap.test}: Categories non-significantly different at the alpha risk of \emph{alpha.total.bootstrap.test} according to the total bootstrap test are linked by a line on the plot. If these links are not required, \emph{alpha.total.bootstrap.test} can be set to 1
}
}
\examples{
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

res=mrCA(dset)

plot(res)
}
