% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sensory.mr.sig.cell.R
\name{plot.sensory.mr.sig.cell}
\alias{plot.sensory.mr.sig.cell}
\title{Plot significant cells}
\usage{
\method{plot}{sensory.mr.sig.cell}(
  x,
  alpha.1 = 0.05,
  alpha.2 = 0,
  choice = "percent.derived.cont",
  col.greater.1 = "green3",
  col.lower.1 = "orangered",
  col.greater.2 = "lightgreen",
  col.lower.2 = "lightsalmon",
  ...
)
}
\arguments{
\item{x}{A list returned by \code{\link[MultiResponseR]{sensory.mr.sig.cell}}}

\item{alpha.1}{The alpha risk to consider the tests as significant}

\item{alpha.2}{The alpha risk to consider the tests as showing a trend. If trends are not to be considered, \emph{alpha.2} can be set to 0 (Default)}

\item{choice}{Which table from \emph{res} should be plotted? Default is percent.derived.cont}

\item{col.greater.1}{The color used to highlight significant positive associations}

\item{col.lower.1}{The color used to highlight significant negative associations}

\item{col.greater.2}{The color used to highlight positive associations showing a trend}

\item{col.lower.2}{The color used to highlight negative associations showing a trend}

\item{...}{further arguments passed to or from other methods}
}
\value{
A table with cells highlighted
}
\description{
This function plots the results coming from \code{\link[MultiResponseR]{sensory.mr.sig.cell}}
}
\examples{
data(milkchoc)

dim.sig=sensory.mr.dimensionality.test(milkchoc)$dim.sig

res=sensory.mr.sig.cell(milkchoc,nbaxes.sig=dim.sig)

plot(res)
}
