% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.mr.sig.cell.R
\name{plt.mr.sig.cell}
\alias{plt.mr.sig.cell}
\title{Plot significant cells}
\usage{
plt.mr.sig.cell(
  res,
  alpha.1 = 0.05,
  alpha.2 = 0.15,
  choice = "percent.derived.cont",
  col.greater.1 = "green3",
  col.lower.1 = "orangered",
  col.greater.2 = "lightgreen",
  col.lower.2 = "lightsalmon"
)
}
\arguments{
\item{res}{A list returned by \code{\link[MultiResponseR]{sensory.mr.sig.cell}} or \code{\link[MultiResponseR]{mr.sig.cell}}}

\item{alpha.1}{The alpha risk to consider the tests as significant}

\item{alpha.2}{The alpha risk to consider the tests as showing a trend. If trends are not to be considered, \emph{alpha.2} can be set to 0}

\item{choice}{Which table from \emph{res} should be plotted? Default is percent.derived.cont}

\item{col.greater.1}{The color used to highlight significant positive associations}

\item{col.lower.1}{The color used to highlight significant negative associations}

\item{col.greater.2}{The color used to highlight positive associations showing a trend}

\item{col.lower.2}{The color used to highlight negative associations showing a trend}
}
\value{
A table with cells highlighted
}
\description{
This function plots the results coming from \code{\link[MultiResponseR]{sensory.mr.sig.cell}} or \code{\link[MultiResponseR]{mr.sig.cell}}
}
\examples{
# non-sensory example
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

parallel::detectCores()

res=mr.sig.cell(dset)

plt.mr.sig.cell(res)

# sensory example
data(milkchoc)

parallel::detectCores()

dim.sig=sensory.mr.dimensionality.test(milkchoc)$dim.sig

res=sensory.mr.sig.cell(milkchoc,nbaxes.sig=dim.sig)

plt.mr.sig.cell(res)
}
