% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.mrCA.R
\name{plt.mrCA}
\alias{plt.mrCA}
\title{Plot factor plan resulting from multiple-response Correspondence Analysis (mrCA)}
\usage{
plt.mrCA(
  res,
  axes = c(1, 2),
  alpha.total.bootstrap.test = 0.05,
  alpha.ellipse = alpha.total.bootstrap.test,
  select.desc.rep = rownames(res$col.coord),
  rev.x = FALSE,
  rev.y = FALSE,
  size.points = 1.4,
  size.lab = 1.4,
  expansion = 1.25,
  title = NULL
)
}
\arguments{
\item{res}{A list returned by \code{\link[MultiResponseR]{sensory.mrCA}} or \code{\link[MultiResponseR]{mrCA}}}

\item{axes}{Which dimensions of the mrCA should be plotted?}

\item{alpha.total.bootstrap.test}{The α risk of the total bootstrap tests. Only useful if the mrCA was computed using \code{\link[MultiResponseR]{sensory.mrCA}} or \code{\link[MultiResponseR]{mrCA}} and ellipse=TRUE. See details}

\item{alpha.ellipse}{The α risk of the confidence ellipses. Only useful if the mrCA was computed using \code{\link[MultiResponseR]{sensory.mrCA}} or \code{\link[MultiResponseR]{mrCA}} and ellipse=TRUE}

\item{select.desc.rep}{A character vector specifying the descriptors/response options to plot. By default, all descriptors/response options are plotted}

\item{rev.x}{Should the horizontal plotted dimension be reversed? Useful in case of map comparisons to align products/categories}

\item{rev.y}{Should the vertical plotted dimension be reversed? Useful in case of map comparisons to align products/categories}

\item{size.points}{The size of the points used to represent the products/categories on the map}

\item{size.lab}{The size of the label on the map}

\item{expansion}{The factor of expansion applied to descriptors/response options coordinates to increase readability}

\item{title}{An optional title to be added to the plot}
}
\value{
An mrCA factor map


}
\description{
This function plots the results coming from \code{\link[MultiResponseR]{sensory.mrCA}} or \code{\link[MultiResponseR]{mrCA}}
}
\details{
\itemize{
\item \strong{alpha.total.bootstrap.test}: Products/categories non-significantly different at the α risk of \emph{alpha.total.bootstrap.test} according to the total bootstrap test are linked by a line on the plot. If these links are not required, \emph{alpha.total.bootstrap.test} can be set to 1
}
}
\examples{
# non-sensory example
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

parallel::detectCores()

res=mrCA(dset)

plt.mrCA(res)

# sensory example
data(milkchoc)

parallel::detectCores()

dim.sig=sensory.mr.dimensionality.test(milkchoc)$dim.sig

res=sensory.mrCA(milkchoc,nbaxes.sig=dim.sig)

plt.mrCA(res)
}
