% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensory.mr.dimensionality.test.R
\name{sensory.mr.dimensionality.test}
\alias{sensory.mr.dimensionality.test}
\title{Multiple-response dimensionality test for sensory data}
\usage{
sensory.mr.dimensionality.test(data, nperm = 2000, alpha = 0.05, ncores = 2)
}
\arguments{
\item{data}{A data.frame of evaluations in rows whose first two columns are factors (subject and product) and subsequent columns are binary numeric or integer, each column being a descriptor}

\item{nperm}{Number of permuted datasets to estimate the distribution of the statistic under the null hypothesis. See details}

\item{alpha}{The alpha risk of the test}

\item{ncores}{Number of cores used to estimate the null distribution. Default is 2. See details}
}
\value{
A list with the following elements:
\describe{
\item{dim.sig}{The number of significant dimensions}
\item{statistics}{Observed multiple-response chi-square statistic of each dimension}
\item{p.values}{P-value of the test of each dimension adjusted for closed testing procedure}
}
}
\description{
Performs a multiple-response dimensionality test as defined in Mahieu, Schlich, Visalli, and Cardot (2020) using random permutations to estimate the null distribution. This test results of the transposition of the dimensionality test introduced in Mahieu, Visalli, and Schlich (2020) from the usual chi-square framework to the modified chi-square framework introduced in Loughin and Scherer (1998). The difference with \code{\link[MultiResponseR]{mr.dimensionality.test}} is that random permutations are performed within subjects rather than along all evaluations
}
\details{
\itemize{
\item \strong{nperm}: The distribution of the statistic under the null hypothesis of no associations between products and descriptors is estimated using \emph{nperm} datasets generated thanks to random permutations of the response vectors along products within subjects.
\item \strong{ncores}: The more cores are added in the process, the faster the results will be obtained. The number of available cores is accessible using \code{\link[parallel]{detectCores}}. The parallel tasks are closed once the \emph{nperm} datasets are generated.
}
}
\examples{
data(milkchoc)

parallel::detectCores()

sensory.mr.dimensionality.test(milkchoc)
}
\references{
Loughin, T. M., & Scherer, P. N. (1998). Testing for Association in Contingency Tables with Multiple Column Responses. Biometrics, 54(2), 630-637.

Mahieu, B., Visalli, M., & Schlich, P. (2020). Accounting for the dimensionality of the dependence in analyses of contingency tables obtained with Check-All-That-Apply and Free-Comment. Food Quality and Preference, 83.

Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2020). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Manuscript submitted for publication.
}
