% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensory.overall.analysis.R
\name{sensory.overall.analysis}
\alias{sensory.overall.analysis}
\title{Overall analysis of multiple-response sensory data using the multiple-response chi-square framework introduced in Mahieu, Schlich, Visalli, and Cardot (2021)}
\usage{
sensory.overall.analysis(
  data,
  nMC = 2000,
  alpha = 0.05,
  cell.two.sided = FALSE
)
}
\arguments{
\item{data}{A data.frame of evaluations in rows whose first two columns are factors (subject and product) and subsequent columns are binary numeric or integer, each column being a descriptor}

\item{nMC}{Number of Monte-Carlo simulations to consider at each step of the overall analysis}

\item{alpha}{The alpha risk to consider at each step of the overall analysis}

\item{cell.two.sided}{Logical. Should the multiple-response hypergeometric tests per cell be two-sided or not? By default, the tests are performed with a one-sided greater alternative hypothesis}
}
\value{
The first MR-CA factor map and the percent.derived.cont table with significant cells highlighted
}
\description{
Successively performs \code{\link[MultiResponseR]{sensory.mr.dimensionality.test}}, \code{\link[MultiResponseR]{sensory.mrCA}} and \code{\link[MultiResponseR]{sensory.mr.sig.cell}}
}
\examples{
data(milkchoc)

sensory.overall.analysis(milkchoc)
}
\references{
Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2021). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Food Quality and Preference, 93.
}
