% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.sig.cell.R
\name{mr.sig.cell}
\alias{mr.sig.cell}
\title{Multiple-response tests per cell}
\usage{
mr.sig.cell(
  data,
  nsample = 2000,
  nbaxes.sig = Inf,
  two.sided = FALSE,
  ncores = 2
)
}
\arguments{
\item{data}{A data.frame of observations in rows whose first column is a factor (the categories) and subsequent columns are binary numeric or integer, each column being a response option}

\item{nsample}{Number of randomly sampled datasets to estimate the distribution of the value under the null hypothesis. See details}

\item{nbaxes.sig}{The number of significant axes retuned by \code{\link[MultiResponseR]{mr.dimensionality.test}}. By default, all axes are considered significant. See details}

\item{two.sided}{Logical. Should the tests be two-sided or not? By default, the tests are performed with a one-sided greater alternative hypothesis}

\item{ncores}{Number of cores used to estimate the null distribution. Default is 2. See details}
}
\value{
A list with the following elements:
\describe{
\item{original.cont}{Observed number of times each category chosen each response option}
\item{percent.cont}{Within each category, percentage of observations where the response options were chosen}
\item{null.cont}{Expected number of times each category chosen each response option under the null hypothesis}
\item{p.values}{P-values of the tests per cell}
\item{derived.cont}{The derived contingency table corresponding to \emph{nbaxes.sig} axes}
\item{percent.derived.cont}{Within each category, percentage of observations where the response options were chosen in the derived contingency table corresponding to \emph{nbaxes.sig} axes}
}
}
\description{
This function performs the tests per cell as defined in Mahieu, Schlich, Visalli, and Cardot (2020) using random hypergeometric samplings to estimate the null distribution. For each cell, a test is performed to investigate if the observed value significantly differs from the expected value under the null hypothesis of no associations between categories and response options. The test performed for each cell results of the adaptation of the chi-square per cell test described in Symoneaux, Galmarini, and Mehinagic (2012) from the usual chi-square framework to the modified chi-square framework introduced in Loughin and Scherer (1998)
}
\details{
\itemize{
\item \strong{nsample}: The distribution of the value under the null hypothesis of no associations between categories and response options is estimated using \emph{nsample} datasets generated thanks to random hypergeometric samplings of the response vectors along observations.
\item \strong{nbaxes.sig}: If \emph{nbaxes.sig} is lower than the total number of axes then the tests are performed on the derived contingency table corresponding to significant axes (Mahieu, Visalli, & Schlich, 2020; Mahieu, Schlich, Visalli, & Cardot, 2020). This table is obtained by reversing mrCA computations on the first \emph{nbaxes.sig} axes.
\item \strong{ncores}: The more cores are added in the process, the faster the results will be obtained. The number of available cores is accessible using \code{\link[parallel]{detectCores}}. The parallel tasks are closed once the \emph{nsample} datasets are generated.
}
}
\examples{
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

parallel::detectCores()

res=mr.sig.cell(dset)

plt.mr.sig.cell(res)
}
\references{
Symoneaux, R., Galmarini, M. V., & Mehinagic, E. (2012). Comment analysis of consumer’s likes and dislikes as an alternative tool to preference mapping. A case study on apples. Food Quality and Preference, 24(1), 59-66.

Loughin, T. M., & Scherer, P. N. (1998). Testing for Association in Contingency Tables with Multiple Column Responses. Biometrics, 54(2), 630-637.

Mahieu, B., Visalli, M., & Schlich, P. (2020). Accounting for the dimensionality of the dependence in analyses of contingency tables obtained with Check-All-That-Apply and Free-Comment. Food Quality and Preference, 83.

Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2020). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Manuscript submitted for publication.
}
