% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiResponseR-package.R
\docType{package}
\name{MultiResponseR-package}
\alias{MultiResponseR}
\alias{MultiResponseR-package}
\title{MultiResponseR: Analysis of multiple-response contingency data}
\description{
This package implements the multiple-response chi-square framework introduced in Mahieu, Schlich, Visalli, and Cardot (2021) for the analysis of contingency data obtained from multiple-response questionnaires. The multiple-response framework notably includes a multiple-response chi-square test and a multiple-response correspondence analysis. Two types of cases are distinguished: a general case where one subject contributes only to one category and the particular case of sensory data (notably Check-All-That-Apply and Free-Comment) where one subject contributes to several categories (products).
}
\references{
Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2021). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Food Quality and Preference, 93.

Loughin, T. M., & Scherer, P. N. (1998). Testing for Association in Contingency Tables with Multiple Column Responses. Biometrics, 54(2), 630-637.
}
\seealso{
\code{\link[MultiResponseR]{mr.chisq.test}}

\code{\link[MultiResponseR]{mr.dimensionality.test}}

\code{\link[MultiResponseR]{mr.sig.cell}}

\code{\link[MultiResponseR]{mrCA}}

\code{\link[MultiResponseR]{plt.mr.sig.cell}}

\code{\link[MultiResponseR]{plt.mrCA}}

\code{\link[MultiResponseR]{sensory.mr.dimensionality.test}}

\code{\link[MultiResponseR]{sensory.mr.sig.cell}}

\code{\link[MultiResponseR]{sensory.mrCA}}

\code{\link[MultiResponseR]{sensory.overall.analysis}}
}
\author{
\strong{Maintainer}: Benjamin Mahieu \email{benjaminmahieu@yahoo.com}

Other contributors:
\itemize{
  \item Michel Visalli \email{michel.visalli@inrae.fr} (Centre des Sciences du Gout et de l'Alimentation, AgroSup Dijon, CNRS, INRAE, Universite Bourgogne Franche-Comte) [contributor]
  \item Herve Cardot \email{herve.cardot@u-bourgogne.fr} (Institut de Mathematiques de Bourgogne, CNRS, Universite Bourgogne Franche-Comte) [contributor, thesis advisor]
  \item Pascal Schlich \email{pascal.schlich@inrae.fr} (Centre des Sciences du Gout et de l'Alimentation, AgroSup Dijon, CNRS, INRAE, Universite Bourgogne Franche-Comte) [thesis advisor]
}

}
\keyword{internal}
