% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.mr.sig.cell.R
\name{plt.mr.sig.cell}
\alias{plt.mr.sig.cell}
\title{Plot significant cells}
\usage{
plt.mr.sig.cell(
  res,
  alpha = 0.075,
  choice = "percent.derived.cont",
  col.greater = "green3",
  col.lower = "orangered"
)
}
\arguments{
\item{res}{A list returned by \code{\link[MultiResponseR]{sensory.mr.sig.cell}} or \code{\link[MultiResponseR]{mr.sig.cell}}}

\item{alpha}{The α risk to consider the tests as significant. Note that the tests can be slightly conservative as suggested by Appendix in Mahieu, Schlich, Visalli, and Cardot (2020)}

\item{choice}{Which table from \emph{res} should be plotted? Default is percent.derived.cont}

\item{col.greater}{The color used to highlight significant positive associations}

\item{col.lower}{The color used to highlight significant negative associations}
}
\value{
A table with significant cells highlighted
}
\description{
This function plots the results coming from \code{\link[MultiResponseR]{sensory.mr.sig.cell}} or \code{\link[MultiResponseR]{mr.sig.cell}}
}
\examples{
# non-sensory example
nb.obs=200
nb.response=5
nb.category=5
vec.category=paste("C",1:nb.category,sep="")
right=matrix(rbinom(nb.response*nb.obs,1,0.25),nb.obs,nb.response)
category=sample(vec.category,nb.obs,replace = TRUE)
dset=cbind.data.frame(category,right)
dset$category=as.factor(dset$category)

parallel::detectCores()

res=mr.sig.cell(dset)

plt.mr.sig.cell(res)

# sensory example
data(milkchoc)

parallel::detectCores()

dim.sig=sensory.mr.dimensionality.test(milkchoc)$dim.sig

res=sensory.mr.sig.cell(milkchoc,nbaxes.sig=dim.sig)

plt.mr.sig.cell(res)
}
\references{
Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2020). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Manuscript submitted for publication.
}
