% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensory.mr.sig.cell.R
\name{sensory.mr.sig.cell}
\alias{sensory.mr.sig.cell}
\title{Multiple-response tests per cell for sensory data}
\usage{
sensory.mr.sig.cell(data, nsample = 2000, nbaxes.sig = Inf, two.sided = FALSE)
}
\arguments{
\item{data}{A data.frame of evaluations in rows whose first two columns are factors (subject and product) and subsequent columns are binary numeric or integer, each column being a descriptor}

\item{nsample}{Number of randomly sampled datasets to estimate the distribution of the value under the null hypothesis. See details}

\item{nbaxes.sig}{The number of significant axes retuned by \code{\link[MultiResponseR]{sensory.mr.dimensionality.test}}. By default, all axes are considered significant. See details}

\item{two.sided}{Logical. Should the tests be two-sided or not? By default, the tests are performed with a one-sided greater alternative hypothesis}
}
\value{
A list with the following elements:
\describe{
\item{original.cont}{Observed number of times each product was described by each descriptor}
\item{percent.cont}{For each product, percentage of evaluations where each descriptor was cited for this product}
\item{null.cont}{Expected number of times each product was described by each descriptor under the null hypothesis}
\item{p.values}{P-values of the tests per cell fdr adjusted by descriptor}
\item{derived.cont}{The derived contingency table corresponding to \emph{nbaxes.sig} axes}
\item{percent.derived.cont}{For each product, percentage of evaluations where each descriptor was cited for this product in the derived contingency table corresponding to \emph{nbaxes.sig} axes}
}
}
\description{
This function performs for each pair of product and descriptor a multiple-response hypergeometric test as defined in Mahieu, Schlich, Visalli, and Cardot (2021) using random hypergeometric samplings to estimate the null distribution. The difference with \code{\link[MultiResponseR]{mr.sig.cell}} is that random hypergeometric samplings are performed within subjects in \code{\link[MultiResponseR]{sensory.mr.sig.cell}}
}
\details{
\itemize{
\item \strong{nsample}: The distribution of the value under the null hypothesis of no associations between products and descriptors is estimated using \emph{nsample} datasets generated thanks to random hypergeometric samplings of the response vectors along products within subjects.
\item \strong{nbaxes.sig}: If \emph{nbaxes.sig} is lower than the total number of axes then the tests are performed on the derived contingency table corresponding to significant axes (Mahieu, Schlich, Visalli, & Cardot, 2021) This table is obtained by using the reconstitution formula of MR-CA on the first \emph{nbaxes.sig} axes.
}
}
\examples{
data(milkchoc)

dim.sig=sensory.mr.dimensionality.test(milkchoc)$dim.sig

res=sensory.mr.sig.cell(milkchoc,nbaxes.sig=dim.sig)

plot(res)
}
\references{
Loughin, T. M., & Scherer, P. N. (1998). Testing for Association in Contingency Tables with Multiple Column Responses. Biometrics, 54(2), 630-637.

Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2021). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Food Quality and Preference, 93.
}
