% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensory.overall.analysis.R
\name{sensory.overall.analysis}
\alias{sensory.overall.analysis}
\title{Overall analysis of multiple-response sensory data using the multiple-response chi-square framework introduced in Mahieu, Schlich, Visalli, and Cardot (2020)}
\usage{
sensory.overall.analysis(
  data,
  nMC = 2000,
  alpha = 0.05,
  cell.two.sided = FALSE,
  ncores = 2
)
}
\arguments{
\item{data}{A data.frame of evaluations in rows whose first two columns are factors (subject and product) and subsequent columns are binary numeric or integer, each column being a descriptor}

\item{nMC}{Number of Monte-Carlo simulations to consider at each step of the overall analysis}

\item{alpha}{The alpha risk to consider at each step of the overall analysis}

\item{cell.two.sided}{Logical. Should the multiple-response tests per cell be two-sided or not? By default, the tests are performed with a one-sided greater alternative hypothesis}

\item{ncores}{Number of cores used in the Monte-Carlo simulations. Default is 2. See details}
}
\value{
The first mrCA factor map and the percent.derived.cont table with significant cells highlighted
}
\description{
Successively performs \code{\link[MultiResponseR]{sensory.mr.dimensionality.test}}, \code{\link[MultiResponseR]{sensory.mrCA}} and \code{\link[MultiResponseR]{sensory.mr.sig.cell}}
}
\details{
\itemize{
\item \strong{ncores}: The more cores are added in the process, the faster the results will be obtained. The number of available cores is accessible using \code{\link[parallel]{detectCores}}. The parallel tasks are closed once the simulations are over.
}
}
\examples{
data(milkchoc)

parallel::detectCores()

sensory.overall.analysis(milkchoc)
}
\references{
Mahieu, B., Schlich, P., Visalli, M., & Cardot, H. (2020). A multiple-response chi-square framework for the analysis of Free-Comment and Check-All-That-Apply data. Manuscript submitted for publication.
}
